--- backend/blog-batch/src/main/kotlin/com/contentria/batch/BatchJobExecutor.kt
@@ -1,7 +1,0 @@
-package com.contentria.batch
-
-interface BatchJobExecutor {
-
-    fun getJobName(): String
-    fun execute()
-}
