--- backend/blog-batch/src/main/kotlin/com/contentria/batch/global/config/AppConfig.kt
@@ -1,8 +1,0 @@
-package com.contentria.batch.global.config
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.context.annotation.Configuration
-
-@EnableConfigurationProperties(AppProperties::class)
-@Configuration
-class AppConfig
