--- backend/blog-batch/src/main/kotlin/com/contentria/batch/global/config/SchedulerConfig.kt
@@ -1,6 +1,0 @@
-package com.contentria.batch.global.config
-
-//@EnableScheduling
-//@Configuration
-//class SchedulerConfig {
-//}
