+++ backend/blog-batch/src/main/resources/.env.sample
@@ -1,0 +1,3 @@
+DATABASE_URL=
+DATABASE_USERNAME=
+DATABASE_PASSWORD=
